Made an couple of Changes
Okay, everything should remain the same.

The Scripts are in the Folders that i've organized into and nothing from
there is changed

What i did:
Added "Main Menu" and Menu script to the UI folder
Adjusted the UI for Gameplay and add hearts to the screen.

No major changes made and it should not be any issues there